Update 02._Discover_Basic_Elements.md
Attempting to replace 98 &nbsp; with spaces.  Scared.  (not all that ewasy when my MarkdownPad 2 and the github editor... don't have REPLACE.  They only have FIND.  The cleanup was done in my Boxer text editor.

In docs... (http://www.html5gamedevs.com/topic/16365-contributing-to-documentation-101/) there is a sentence...

When done, just select "Create a new branch for this commit and start a pull request" 

That was not seen anywhere in the github editor.

But I DO see Propose file change.  This comment is being typed there.  

Sorry if I made a mistake.  Do not commit this IF there is some reason for all the &nbsp; that were found in this file.  In other words, ignore this IF the &nbsp; are there for some reason.